requirements.txt
flask
bcrypt
pyotp
python-dotenv
qrcode
pillow
jwt
PyJWT
